% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{convert_to_RBP}
\alias{convert_to_RBP}
\title{This function takes the information on particular gene locations and creates a file to search for RBPs on RBPMap}
\usage{
convert_to_RBP(
  chr.vector,
  start.vector,
  end.vector,
  strand.vector,
  bprange = 100,
  path.to.output = "output.txt"
)
}
\arguments{
\item{chr.vector}{Chromosome locations}

\item{start.vector}{Start positions}

\item{end.vector}{End positions}

\item{strand.vector}{Strand positions}

\item{bprange}{BP range}
}
\description{
This function takes the information on particular gene locations and creates a file to search for RBPs on RBPMap
}
