% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{inputRBP}
\alias{inputRBP}
\title{This function takes the summary results from RBPMap and converts it to a list of lists}
\usage{
inputRBP(path.to.file)
}
\arguments{
\item{path.to.file}{The path of the RBPMap text file}
}
\value{
the resulting list of lists summarizing RBPMap results
}
\description{
This function takes the summary results from RBPMap and converts it to a list of lists
}
