% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{sampleRBP}
\alias{sampleRBP}
\title{This function samples a list from an RBPMap results file}
\usage{
sampleRBP(path.to.file, samplecount)
}
\arguments{
\item{path.to.file}{The path of the RBPMap text file}

\item{samplecount}{Number of samples needed}
}
\value{
the resulting list of lists summarizing RBPMap results
}
\description{
This function samples a list from an RBPMap results file
}
